requirements.txt
mcp>=1.3.0
gradio>=4.0.0
loguru>=0.7.0
pydantic>=2.6.0
aiohttp>=3.11.0
nest-asyncio>=1.6.0
packaging>=24.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
pillow>=10.0.0  # 画像処理用

# 開発用依存関係
pytest>=8.0.0
pytest-asyncio>=0.25.0
mypy>=1.8.0
black>=24.0.0
isort>=5.12.0
pre-commit>=3.5.0 